potential fix

+++ io.openems.edge.meter.api/src/io/openems/edge/meter/api/Meter.java
@@ -52,7 +52,7 @@ public interface Meter extends OpenemsComponent {
         /**
          * Generic Meter Reading. Each Meter decides on it's own what to put here.
          */
-        METER_READING(Doc.of(OpenemsType.LONG).accessMode(AccessMode.READ_ONLY));
+        METER_READING(Doc.of(OpenemsType.DOUBLE).accessMode(AccessMode.READ_ONLY));
 
 
         private final Doc doc;
